[IR] Remove custom error messages in IR declaration implementations
Specifically, remove custom assert on uninitialized
'parent' and 'returnType' properties.

This is to remove custom logic from those classes,
so that source generator will be able to generate
them in a straight-forward way.
Otherwise, it would need to be thought about those special cases, which
is better to be avoided.

As for the 'parent' property, it is also a small step for future work,
which likely involves making it nullable.

Unfortunately, this change causes a loss of some details from exception
messages.

^KT-65773 In Progress